+++ app/models/transfer/task_list.rb
@@ -30,6 +30,7 @@ class Transfer::TaskList < ::BaseTaskList
         identifier: :get_ready_for_opening,
         tasks: [
           Transfer::Task::ConditionsMetTaskForm,
+          Transfer::Task::ConfirmIncomingTrustHasCompletedAllActionsTaskForm,
           Transfer::Task::RpaPolicyTaskForm
         ]
       }