+++ db/migrate/20230919115055_add_transfer_confirm_incoming_trust_has_completed_all_actions_task_attributes.rb
@@ -1,0 +1,6 @@
+class AddTransferConfirmIncomingTrustHasCompletedAllActionsTaskAttributes < ActiveRecord::Migration[7.0]
+  def change
+    add_column :transfer_tasks_data, :confirm_incoming_trust_has_completed_all_actions_emailed, :boolean
+    add_column :transfer_tasks_data, :confirm_incoming_trust_has_completed_all_actions_saved, :boolean
+  end
+end